% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSPARQLanything-package.R
\docType{package}
\name{rSPARQLanything-package}
\alias{rSPARQLanything}
\alias{rSPARQLanything-package}
\title{rSPARQLanything: Query Any Data Resource with SPARQL in R}
\description{
rSPARQLanything provides an interface to SPARQL Anything, a system for Semantic Web re-engineering that allows users to query anything with SPARQL. It uses a single generic abstraction for all data source formats called Facade-X. This meta-model used by SPARQL Anything transformers to generate RDF data (a widely used data representation model that forms the cornerstone of the Semantic Web) from diverse data sources. Intuitively, Facade-X uses a subset of RDF as a general approach to represent the source content as-it-is but in RDF. The package interface takes inspiration from the Python 'pysparql-anything' library.
}
\author{
\strong{Maintainer}: Daniel Antal \email{daniel.antal@dataobservatory.eu} (\href{https://orcid.org/0000-0001-7513-6760}{ORCID})

}
\keyword{internal}
