% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_anything.R
\name{sparql_anything}
\alias{sparql_anything}
\title{SPARQL-Anything wrapper}
\usage{
sparql_anything(
  query,
  output_file = NULL,
  append = FALSE,
  explain = FALSE,
  format = "TEXT",
  strategy = "triple_filtering",
  output_pattern = NULL,
  values = NULL,
  configuration = NULL
)
}
\arguments{
\item{query}{A SPARQL query as character string.}

\item{output_file}{An optional output file}

\item{append}{If the output file should be appended, defaults to \code{FALSE}.}

\item{explain}{Explain query execution, defaults to \code{FALSE}.}

\item{format}{Defaults to \code{"TEXT"}, \code{"JSON"}, \code{"XML"},
\code{"CSV"}, \code{"TEXT"}, \code{"TTL"}, \code{"NT"}, \code{"NQ"}.}

\item{strategy}{Defaults to \code{"triple_filtering"}.}

\item{output_pattern}{Defaults to \code{NULL}.}

\item{values}{Defaults to \code{NULL}.}

\item{configuration}{Defaults to \code{NULL}.}
}
\value{
A tabular text, as the result of the SPARQL in according to the 
\code{format} parameter.
}
\description{
SPARQL-Anything wrapper
}
\examples{
Not yet implemented.
sparql_anything()
}
